pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9,<3.10"
beautifulsoup4 = "4.12.3"
everypolitician = "0.0.13"
lxml = "5.2.1"
python-dateutil = "2.9.0"
requests = { version = "2.32.3", extras = ["security"] }
requests-cache = "0.4.13"
Click = "^8.0"
click-log = "0.3.2"
mysoc-validator = "^1.2.0"
tqdm = "^4.67.1"
jinja2 = "^3.1.4"
html5lib = "^1.1"
pandas = "^2.2.3"
pydantic-settings = "^2.8.1"
atproto = "^0.0.59"
pyarrow = "^19.0.1"
backports-strenum = "^1.3.1"
blis = "1.2.0"
spacy = "^3.8.7"
pydantic-ai = "0.8.0"
jsonref = "^1.1.0"
duckdb = "^1.4.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"

[tool.ruff]

extend-include = ["scripts/*"]
extend-exclude = [
  'scripts/.gitignore',
  'scripts/2016_data_update/README.txt',
  'scripts/config.pm.incvs',
  'scripts/consts',
  'scripts/crontab',
  'scripts/dailyupdate',
  'scripts/dataupdate',
  'scripts/datadotparl/mp-party-check',
  'scripts/datadotparl/one-off-add-pims-ids',
  'scripts/datadotparl/one-off-sync-lord-parties',
  'scripts/dircmp',
  'scripts/divisionextractor.pl',
  'scripts/morningupdate',
  'scripts/ni-format-revert',
  'scripts/ni_membership.php',
  'scripts/one-off-move-names-to-persons',
  'scripts/other-sites-update',
  'scripts/updatedaterange-parse',
  'scripts/updatedaterange-scrape',
  'scripts/weeklyupdate',
  'scripts/ynmp/tests.txt',
  'scripts/info-tasks'
]


[tool.ruff.lint]
select = [
  "E",
  # flake8
  "F",
  # isort
  "I",
]
ignore = [
  # line too long, sorted with formatter where it can be
  "E501",
]


[tool.ruff.lint.isort]
known-first-party = ["hub"]
section-order = [
  "future",
  "standard-library",
  "django",
  "third-party",
  "first-party",
  "local-folder",
]

[tool.ruff.lint.isort.sections]
django = ["django"]
